Tb signas now added to test data seed; fixed numeric->integer; optional tbsigna belongs_to score

+++ db/migrate/20190315022416_create_truthbooksections.rb
@@ -3,8 +3,8 @@ class CreateTruthbooksections < ActiveRecord::Migration[5.1]
     create_table :truthbooksections do |t|
       t.string :unit
       t.string :section
-      t.numeric :sort
-      t.numeric :truthbook_id
+      t.integer :sort
+      t.integer :truthbook_id
 
       t.timestamps
     end
